FUN-6: Fix issues reported by sonar.

+++ src/main/java/com/hacker/rank/math/medium/Sticks.java
@@ -68,7 +68,7 @@ public class Sticks {
 
         for (long i = factor + 1L; i <= Math.sqrt(bar); i += factor) {
             while (bar % i == 0) {
-                factors.add((long) factor);
+                factors.add(factor);
                 bar = bar / i;
             }
         }
